# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/chat/sessions" : {
      "get" : {
        "tags" : [ "chat-session-controller" ],
        "summary" : "Get the chat session for a specific order",
        "operationId" : "getSessionByOrderId",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseChatSessionResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "chat-session-controller" ],
        "summary" : "Create or retrieve a chat session for an order",
        "operationId" : "createOrGetSession",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseChatSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/sessions/{sessionId}/upload-image" : {
      "post" : {
        "tags" : [ "chat-image-upload-controller" ],
        "operationId" : "uploadImage",
        "parameters" : [ {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "file" ],
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseUploadResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/sessions/{sessionId}/upload-audio" : {
      "post" : {
        "tags" : [ "chat-audio-upload-controller" ],
        "operationId" : "uploadAudio",
        "parameters" : [ {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "file" ],
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseUploadResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/sessions/{sessionId}/participants" : {
      "post" : {
        "tags" : [ "chat-session-controller" ],
        "summary" : "Synchronize an existing order session with its authoritative roster",
        "operationId" : "addParticipant",
        "parameters" : [ {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseChatSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/webrtc/ice-servers" : {
      "get" : {
        "tags" : [ "turn-credential-controller" ],
        "operationId" : "getIceServers",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TurnCredentialsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/sessions/{sessionId}/messages" : {
      "get" : {
        "tags" : [ "chat-session-controller" ],
        "summary" : "Get paginated message history for a session",
        "operationId" : "getMessages",
        "parameters" : [ {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseDtoChatMessageDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "CreateSessionRequest" : {
        "required" : [ "orderId" ],
        "type" : "object",
        "properties" : {
          "orderId" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseChatSessionResponse" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ChatSessionResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ChatSessionResponse" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sessionType" : {
            "type" : "string"
          },
          "referenceId" : {
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "participants" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ParticipantDto"
            }
          }
        }
      },
      "ParticipantDto" : {
        "required" : [ "entityType", "userId" ],
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "entityType" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseUploadResponseDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/UploadResponseDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "UploadResponseDto" : {
        "type" : "object",
        "properties" : {
          "url" : {
            "type" : "string"
          },
          "messageId" : {
            "type" : "string"
          }
        }
      },
      "IceServer" : {
        "required" : [ "urls" ],
        "type" : "object",
        "properties" : {
          "urls" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "credential" : {
            "type" : "string"
          }
        }
      },
      "TurnCredentialsResponse" : {
        "required" : [ "iceServers" ],
        "type" : "object",
        "properties" : {
          "iceServers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IceServer"
            }
          }
        }
      },
      "ApiResponsePageResponseDtoChatMessageDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseDtoChatMessageDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ChatMessageDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sessionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "senderId" : {
            "type" : "string"
          },
          "senderName" : {
            "type" : "string"
          },
          "senderType" : {
            "type" : "string"
          },
          "messageType" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "imageUrl" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageResponseDtoChatMessageDto" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ChatMessageDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}