# src/main/java/com/fooddelivery/chat/controller/ChatImageUploadController.java
package com.fooddelivery.chat.controller;

import com.fooddelivery.chat.dto.ChatMessageDto;
import com.fooddelivery.chat.service.ChatEventBroadcaster;
import com.fooddelivery.chat.service.ChatMessageService;
import com.fooddelivery.chat.service.ChatSessionAccessService;
import com.fooddelivery.chat.service.ChatSessionService;
import com.fooddelivery.common.service.CloudflareR2Service;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import java.util.Map;
import java.util.UUID;

/**
 * REST controller for uploading images in chat sessions.
 * Images are uploaded to Cloudflare R2 with the folder structure: chat/{sessionId}/
 * After upload, an IMAGE-type message is auto-saved and broadcast to all participants.
 */
@RestController
@RequestMapping("/api/v1/chat")
@lombok.extern.slf4j.Slf4j
public class ChatImageUploadController {
    @java.lang.SuppressWarnings("all")

    private final CloudflareR2Service cloudflareR2Service;
    private final ChatSessionService sessionService;
    private final ChatSessionAccessService accessService;
    private final ChatMessageService messageService;
    private final ChatEventBroadcaster chatEventBroadcaster;
    private static final long MAX_FILE_SIZE = 5 * 1024 * 1024; // 5 MB
    private static final int MAX_IMAGE_DIMENSION = 5000; // pixels

    @org.springframework.security.access.prepost.PreAuthorize("isAuthenticated()")
    @PostMapping("/sessions/{sessionId}/upload-image")
    public ResponseEntity<com.fooddelivery.common.dto.ApiResponse<com.fooddelivery.chat.dto.UploadResponseDto>> uploadImage(@PathVariable UUID sessionId, @RequestParam("file") MultipartFile file, Authentication authentication) {
        // 1. Authentication check
        String userId = authentication != null ? authentication.getName() : null;
        if (userId == null) {
            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(com.fooddelivery.common.dto.ApiResponse.error("Authentication required"));
        }
        // 2. Authorization check — must be a participant
        if (!accessService.canAccessSession(sessionId, authentication)) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(com.fooddelivery.common.dto.ApiResponse.error("Access Denied: Not a participant of this chat session"));
        }
        // 3. Validate file presence
        if (file == null || file.isEmpty()) {
            return ResponseEntity.badRequest().body(com.fooddelivery.common.dto.ApiResponse.error("A file must be provided"));
        }
        // 4. Validate content type
        String contentType = file.getContentType();
        if (contentType == null || !contentType.startsWith("image/")) {
            return ResponseEntity.badRequest().body(com.fooddelivery.common.dto.ApiResponse.error("Only image files are allowed"));
        }
        // 5. Validate file size
        if (file.getSize() > MAX_FILE_SIZE) {
            return ResponseEntity.badRequest().body(com.fooddelivery.common.dto.ApiResponse.error("File size exceeds 5MB limit"));
        }
        // 6. Check image count limit (max 4 per session per user)
        long currentImageCount = messageService.countImagesInSessionByUser(sessionId, userId);
        if (currentImageCount >= 4) {
            log.warn("Upload rejected: User {} in Session {} reached the maximum limit of 4 images. Current count: {}", userId, sessionId, currentImageCount);
            return ResponseEntity.badRequest().body(com.fooddelivery.common.dto.ApiResponse.error("Maximum limit of 4 images per user reached for this session"));
        }
        try {
            byte[] imageBytes = file.getBytes();
            // 6. Validate image dimensions (prevent decompression bomb DoS)
            validateImageDimensions(imageBytes);
            // 7. Determine and sanitize file extension
            String extension = ".jpg";
            String originalFilename = file.getOriginalFilename();
            if (originalFilename != null && originalFilename.contains(".")) {
                String extractedExt = originalFilename.substring(originalFilename.lastIndexOf(".")).toLowerCase();
                if (extractedExt.matches("\\.(jpg|jpeg|png|gif|webp)")) {
                    extension = extractedExt;
                }
            }
            // 8. Upload to Cloudflare R2 with folder = chat/{sessionId}
            String folder = "chat/" + sessionId.toString();
            String fileName = UUID.randomUUID().toString() + extension;
            String publicUrl = cloudflareR2Service.uploadImage(imageBytes, folder, fileName, contentType);
            log.info("Chat image uploaded: session={}, user={}, url={}", sessionId, userId, publicUrl);
            // 9. Auto-save an IMAGE message and broadcast to all subscribers
            ChatMessageDto savedMessage = accessService.isSupportModerator(authentication)
                    ? messageService.saveSupportModeratorMessage(sessionId, userId, publicUrl, "IMAGE")
                    : messageService.saveMessage(sessionId, userId, publicUrl, "IMAGE");
            // Broadcast the image message via STOMP
            chatEventBroadcaster.broadcastMessage(sessionId, savedMessage);
            return ResponseEntity.ok(com.fooddelivery.common.dto.ApiResponse.success(
                com.fooddelivery.chat.dto.UploadResponseDto.builder().url(publicUrl).messageId(savedMessage.getId().toString()).build(),
                "Image uploaded successfully"));
        } catch (Exception e) {
            log.error("Failed to upload chat image for session {}", sessionId, e);
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(com.fooddelivery.common.dto.ApiResponse.error("Failed to upload image: " + e.getMessage()));
        }
    }

    /**
     * Validates image dimensions to prevent decompression bomb attacks.
     * A tiny compressed file could decompress to a massive bitmap in memory.
     */
    private void validateImageDimensions(byte[] imageBytes) throws Exception {
        try (
            java.io.ByteArrayInputStream bais = new java.io.ByteArrayInputStream(imageBytes);
            javax.imageio.stream.ImageInputStream iis = javax.imageio.ImageIO.createImageInputStream(bais)) {
            java.util.Iterator<javax.imageio.ImageReader> readers = javax.imageio.ImageIO.getImageReaders(iis);
            if (readers.hasNext()) {
                javax.imageio.ImageReader reader = readers.next();
                try {
                    reader.setInput(iis, true);
                    int width = reader.getWidth(0);
                    int height = reader.getHeight(0);
                    if (width > MAX_IMAGE_DIMENSION || height > MAX_IMAGE_DIMENSION) {
                        throw new Exception("Image dimensions (" + width + "x" + height + ") exceed the " + MAX_IMAGE_DIMENSION + "x" + MAX_IMAGE_DIMENSION + " safety limit.");
                    }
                } finally {
                    reader.dispose();
                }
            } else {
                throw new Exception("Unsupported or invalid image format.");
            }
        }
    }

    @java.lang.SuppressWarnings("all")
    public ChatImageUploadController(final CloudflareR2Service cloudflareR2Service, final ChatSessionService sessionService, final ChatSessionAccessService accessService, final ChatMessageService messageService, final ChatEventBroadcaster chatEventBroadcaster) {
        this.cloudflareR2Service = cloudflareR2Service;
        this.sessionService = sessionService;
        this.accessService = accessService;
        this.messageService = messageService;
        this.chatEventBroadcaster = chatEventBroadcaster;
    }
}